(#58) Moved Camera class in a new file : Camera.py
